The "do not reply me" setting has been added to the talker.

+++ src/main/java/org/telegram/bot/commands/setters/TalkerSetter.java
@@ -1,11 +1,13 @@
 package org.telegram.bot.commands.setters;
 
+import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.telegram.bot.commands.Set;
 import org.telegram.bot.domain.entities.Chat;
 import org.telegram.bot.domain.entities.TalkerDegree;
+import org.telegram.bot.domain.entities.TalkerUserSettings;
 import org.telegram.bot.domain.entities.User;
 import org.telegram.bot.domain.model.request.BotRequest;
 import org.telegram.bot.domain.model.request.Message;
@@ -15,12 +17,8 @@ import org.telegram.bot.enums.BotSpeechTag;
 import org.telegram.bot.enums.Emoji;
 import org.telegram.bot.enums.FormattingStyle;
 import org.telegram.bot.exception.BotException;
-import org.telegram.bot.services.CommandWaitingService;
-import org.telegram.bot.services.InternationalizationService;
-import org.telegram.bot.services.SpeechService;
-import org.telegram.bot.services.TalkerDegreeService;
+import org.telegram.bot.services.*;
 
-import jakarta.annotation.PostConstruct;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
@@ -39,19 +37,25 @@ public class TalkerSetter implements Setter<BotResponse> {
     private static final String CALLBACK_SET_TALKER_COMMAND = CALLBACK_COMMAND + EMPTY_TALKER_COMMAND;
     private static final String SET_IDLE_MINUTES_COMMAND = EMPTY_TALKER_COMMAND + "m";
     private static final String CALLBACK_SET_IDLE_MINUTES_COMMAND = CALLBACK_COMMAND + SET_IDLE_MINUTES_COMMAND;
+    private static final String SET_DO_NOT_REPLY_COMMAND = EMPTY_TALKER_COMMAND + "d";
+    private static final String CALLBACK_SET_DO_NOT_REPLY_COMMAND = CALLBACK_COMMAND + SET_DO_NOT_REPLY_COMMAND;
 
     private final java.util.Set<String> emptyTalkerCommands = new HashSet<>();
     private final java.util.Set<String> setIdleCommands = new HashSet<>();
+    private final java.util.Set<String> setDoNotReplyMeCommands = new HashSet<>();
 
     private final TalkerDegreeService talkerDegreeService;
     private final CommandWaitingService commandWaitingService;
     private final SpeechService speechService;
     private final InternationalizationService internationalizationService;
+    private final UserService userService;
+    private final TalkerUserSettingsService talkerUserSettingsService;
 
     @PostConstruct
     private void postConstruct() {
         emptyTalkerCommands.addAll(internationalizationService.getAllTranslations("setter.talker.emptycommand"));
         setIdleCommands.addAll(internationalizationService.internationalize(SET_IDLE_MINUTES_COMMAND));
+        setDoNotReplyMeCommands.addAll(internationalizationService.internationalize(SET_DO_NOT_REPLY_COMMAND));
     }
 
     @Override
@@ -61,7 +65,7 @@ public class TalkerSetter implements Setter<BotResponse> {
 
     @Override
     public AccessLevel getAccessLevel() {
-        return AccessLevel.MODERATOR;
+        return AccessLevel.FAMILIAR;
     }
 
     @Override
@@ -73,26 +77,32 @@ public class TalkerSetter implements Setter<BotResponse> {
 
         if (message.isCallback()) {
             if (emptyTalkerCommands.contains(lowerCaseCommandText)) {
-                return getTalkerSetterWithKeyboard(message, chat, false);
+                return getTalkerSetterWithKeyboard(message, chat, user, false);
             } else if (containsStartWith(setIdleCommands, lowerCaseCommandText)) {
                 return setIdleMinutes(message, chat, user, commandText);
+            } else if (containsStartWith(setDoNotReplyMeCommands, lowerCaseCommandText)) {
+                return setDoNotReplyMeByCallback(message, chat, user, commandText);
             } else if (containsStartWith(emptyTalkerCommands, lowerCaseCommandText)) {
-                return selectTalkerDegreeByCallback(message, chat, commandText);
+                return selectTalkerDegreeByCallback(message, chat, user, commandText);
             }
         }
 
         if (emptyTalkerCommands.contains(lowerCaseCommandText)) {
-            return getTalkerSetterWithKeyboard(message, chat, true);
+            return getTalkerSetterWithKeyboard(message, chat, user, true);
         } else if (containsStartWith(setIdleCommands, lowerCaseCommandText)) {
             return setIdleMinutes(message, chat, user, commandText);
         } else if (containsStartWith(emptyTalkerCommands, lowerCaseCommandText)) {
-            return selectTalkerDegree(message, chat, commandText);
+            return selectTalkerDegree(message, chat, user, commandText);
         } else {
             throw new BotException(speechService.getRandomMessageByTag(BotSpeechTag.WRONG_INPUT));
         }
     }
 
     private BotResponse setIdleMinutes(Message message, Chat chat, User user, String command) {
+        if (!userService.isUserHaveAccessForCommand(user, AccessLevel.MODERATOR.getValue())) {
+            throw new BotException(speechService.getRandomMessageByTag(BotSpeechTag.NO_ACCESS));
+        }
+
         commandWaitingService.remove(chat, user);
 
         String setIdleMinutesCommand = getLocalizedCommand(command, SET_IDLE_MINUTES_COMMAND);
@@ -121,7 +131,7 @@ public class TalkerSetter implements Setter<BotResponse> {
                 .setResponseSettings(FormattingStyle.HTML);
     }
 
-    private BotResponse getTalkerSetterWithKeyboard(Message message, Chat chat, boolean newMessage) {
+    private BotResponse getTalkerSetterWithKeyboard(Message message, Chat chat, User user, boolean newMessage) {
         log.debug("Request to get talker setter for chat {}", chat.getChatId());
 
         TalkerDegree talkerDegree = talkerDegreeService.get(chat.getChatId());
@@ -133,21 +143,25 @@ public class TalkerSetter implements Setter<BotResponse> {
         if (newMessage) {
             return new TextResponse(message)
                     .setText(responseText)
-                    .setKeyboard(prepareKeyboardWithDegreeButtons())
+                    .setKeyboard(prepareKeyboardWithDegreeButtons(user))
                     .setResponseSettings(FormattingStyle.HTML);
         }
 
         return new EditResponse(message)
                 .setText(responseText)
-                .setKeyboard(prepareKeyboardWithDegreeButtons())
+                .setKeyboard(prepareKeyboardWithDegreeButtons(user))
                 .setResponseSettings(FormattingStyle.HTML);
     }
 
-    private BotResponse selectTalkerDegreeByCallback(Message message, Chat chat, String command) {
+    private BotResponse selectTalkerDegreeByCallback(Message message, Chat chat, User user, String command) {
+        if (!userService.isUserHaveAccessForCommand(user, AccessLevel.MODERATOR.getValue())) {
+            throw new BotException(speechService.getRandomMessageByTag(BotSpeechTag.NO_ACCESS));
+        }
+
         String emptyTalkerCommand = getLocalizedCommand(command, EMPTY_TALKER_COMMAND);
 
         if (command.equals(emptyTalkerCommand)) {
-            return getTalkerSetterWithKeyboard(message, chat, false);
+            return getTalkerSetterWithKeyboard(message, chat, user, false);
         }
 
         int degree;
@@ -166,15 +180,19 @@ public class TalkerSetter implements Setter<BotResponse> {
         TalkerDegree talkerDegree = talkerDegreeService.get(chat.getChatId());
         talkerDegreeService.save(talkerDegree.setDegree(degree));
 
-        return getTalkerSetterWithKeyboard(message, chat, false);
+        return getTalkerSetterWithKeyboard(message, chat, user, false);
     }
 
-    private BotResponse selectTalkerDegree(Message message, Chat chat, String command) {
+    private BotResponse selectTalkerDegree(Message message, Chat chat, User user, String command) {
+        if (!userService.isUserHaveAccessForCommand(user, AccessLevel.MODERATOR.getValue())) {
+            throw new BotException(speechService.getRandomMessageByTag(BotSpeechTag.NO_ACCESS));
+        }
+
         String emptyTalkerCommand = getLocalizedCommand(command, EMPTY_TALKER_COMMAND);
         log.debug("Request to select talker degree");
 
         if (command.equals(emptyTalkerCommand)) {
-            return getTalkerSetterWithKeyboard(message, chat, true);
+            return getTalkerSetterWithKeyboard(message, chat, user, true);
         }
 
         int degree;
@@ -197,7 +215,31 @@ public class TalkerSetter implements Setter<BotResponse> {
                 .setText(speechService.getRandomMessageByTag(BotSpeechTag.SAVED));
     }
 
-    private Keyboard prepareKeyboardWithDegreeButtons() {
+    private BotResponse setDoNotReplyMeByCallback(Message message, Chat chat, User user, String command) {
+        String emptyTalkerCommand = getLocalizedCommand(command, EMPTY_TALKER_COMMAND);
+
+        if (command.equals(emptyTalkerCommand)) {
+            return getTalkerSetterWithKeyboard(message, chat, user, false);
+        }
+
+        TalkerUserSettings talkerUserSettings = talkerUserSettingsService.get(user);
+        if (talkerUserSettings != null) {
+            talkerUserSettingsService.remove(talkerUserSettings);
+        } else {
+            talkerUserSettingsService.save(new TalkerUserSettings().setChat(chat).setUser(user).setDoNotReply(true));
+        }
+
+        return getTalkerSetterWithKeyboard(message, chat, user, false);
+    }
+
+    private Keyboard prepareKeyboardWithDegreeButtons(User user) {
+        String doNotReplyButtonEmoji;
+        if (talkerUserSettingsService.doNotReply(user)) {
+            doNotReplyButtonEmoji = Emoji.CHECK_MARK_BUTTON.getSymbol();
+        } else {
+            doNotReplyButtonEmoji = Emoji.NO_ENTRY.getSymbol();
+        }
+
         return new Keyboard(List.of(
                 Stream.of(0, 5, 10, 15, 20, 25, 30)
                         .map(value -> new KeyboardButton()
@@ -217,6 +259,9 @@ public class TalkerSetter implements Setter<BotResponse> {
                 List.of(new KeyboardButton()
                         .setName(Emoji.HOURGLASS_NOT_DONE.getSymbol() + "${setter.talker.button.downtime}")
                         .setCallback(CALLBACK_SET_IDLE_MINUTES_COMMAND)),
+                List.of(new KeyboardButton()
+                        .setName(doNotReplyButtonEmoji + "${setter.talker.button.donotreplyme}")
+                        .setCallback(CALLBACK_SET_DO_NOT_REPLY_COMMAND)),
                 List.of(new KeyboardButton()
                         .setName(Emoji.BACK.getSymbol() + "${setter.talker.button.settings}")
                         .setCallback(CALLBACK_COMMAND + "back"))));
